requirements.txt
transformers
Pillow
pandas
torch
scipy